Refactor: Moved handling of malicious files over to fileHandler
