+++ src/Utils/fileHandler.c
@@ -1,7 +1,176 @@
+#include <stdio.h>
+#include <errno.h>
+#include <sys/stat.h>
+#include <linux/limits.h>
+
 #include "Crypto/fingerprint.h"
-#include "stdio.h"
+#include "Utils/logger.h"
 
 #include "fileHandler.h"
+
+// Path to the whitelist file
+#define WHITELIST_PATH "/usr/local/share/pproc/whitelist.txt"
+
+int handle_file(const char* target_file) {
+    log_message(LL_WARNING, "Malicious file detected: %s", target_file);
+    
+    struct stat file_stat;
+    unsigned int file_permissions;
+
+    if (stat(target_file, &file_stat) == -1) {
+        log_message(LL_ERROR, "Could not stat file %s", target_file);
+        return -1;
+    }
+
+    file_permissions = file_stat.st_mode;
+
+    if (chmod(target_file, S_IRUSR | S_IRGRP | S_IROTH) == -1) {
+        log_message(LL_ERROR, "Failed to change file permissions");
+        return 1;
+    }
+
+    printf("\nPossible malicious file detected: %s", target_file);
+    if (get_user_input("\nWould you like to remove the file Y/N:") == 1) {
+        log_message(LL_INFO, "Removing file %s", target_file);
+        if (remove(target_file) != 0) {
+            log_message(LL_ERROR, "Failed to remove file: %s", strerror(errno));
+        }
+    } else {
+        if (get_user_input("\nWould you like to quarantine the file Y/N:") == 1) {
+            struct stat st = {0};
+            if (stat("/usr/local/share/pproc/quarantine", &st) == -1) {
+                if (mkdir("/usr/local/share/pproc/quarantine", 0777) == -1) {
+                    log_message(LL_ERROR, "Failed to create quarantine directory: %s", strerror(errno));
+                    chmod(target_file, file_permissions);
+                }
+                chmod("/usr/local/share/pproc/quarantine", 0777);
+            }
+
+            const char *filename = strrchr(target_file, '/');
+            filename = filename ? filename + 1 : target_file;
+            
+            char quarantine_path[PATH_MAX];
+            snprintf(quarantine_path, sizeof(quarantine_path), 
+                    "/usr/local/share/pproc/quarantine/%s", filename);
+
+            // Open source file
+            FILE *src = fopen(target_file, "rb");
+            if (!src) {
+                log_message(LL_ERROR, "Failed to open source file for quarantine: %s", strerror(errno));
+                chmod(target_file, file_permissions);
+            }
+
+            // Open destination file with explicit permissions
+            FILE *dst = fopen(quarantine_path, "wb");
+            if (!dst) {
+                log_message(LL_ERROR, "Failed to open quarantine destination: %s", strerror(errno));
+                fclose(src);
+                chmod(target_file, file_permissions);
+            }
+
+            // Set permissions on the quarantine file
+            chmod(quarantine_path, 0666);
+
+            // Copy file contents
+            char buf[8192];
+            size_t size;
+            int copy_success = 1;
+            
+            while ((size = fread(buf, 1, sizeof(buf), src)) > 0) {
+                if (fwrite(buf, 1, size, dst) != size) {
+                    copy_success = 0;
+                    break;
+                }
+            }
+
+            fclose(src);
+            fclose(dst);
+
+            if (copy_success) {
+                // Log the original path and permissions before removing the file
+                FILE *quarantine_log = fopen("/usr/local/share/pproc/quarantine_log.txt", "a");
+                if (quarantine_log) {
+                    fprintf(quarantine_log, "%s %o\n", target_file, file_permissions);
+                    fclose(quarantine_log);
+                } else {
+                    log_message(LL_ERROR, "Failed to open quarantine log file");
+                }
+
+                if (remove(target_file) == 0) {
+                    log_message(LL_INFO, "File quarantined successfully: %s", filename);
+                } else {
+                    log_message(LL_ERROR, "Failed to remove original file after quarantine: %s", strerror(errno));
+                    chmod(target_file, file_permissions);
+                    remove(quarantine_path);
+                }
+            } else {
+                log_message(LL_ERROR, "Failed to copy file to quarantine: %s", strerror(errno));
+                chmod(target_file, file_permissions);
+                remove(quarantine_path);
+            }
+        } else {
+            if (get_user_input("\nWould you like to add this file to the whitelist Y/N:") == 1) {
+                add_to_whitelist(target_file);
+            }
+            chmod(target_file, file_permissions);
+        }
+    }
+}
+
+
+int is_whitelisted(const char* target_path) {
+    int result = 0;
+    
+    FILE* whitelist_file = fopen(WHITELIST_PATH, "r");
+    if (whitelist_file == NULL) {
+        log_message(LL_DEBUG, "Whitelist file not found, assuming file is not whitelisted");
+        return 1;
+    }
+
+    char line[PATH_MAX];
+    while (fgets(line, sizeof(line), whitelist_file)) {
+        // Remove newline character
+        line[strcspn(line, "\n")] = 0;
+        
+        if (strcmp(line, target_path) == 0) {
+            result = 1;
+        }
+    }
+
+    fclose(whitelist_file);
+    return result;
+}
+
+int add_to_whitelist(const char* file_path) {    
+    char command[PATH_MAX * 2];
+    
+    // Create directory and set permissions
+    snprintf(command, sizeof(command), 
+        "sudo sh -c '"
+        "mkdir -p /usr/local/share/pproc && "
+        "touch /usr/local/share/pproc/whitelist.txt && "
+        "chmod 777 /usr/local/share/pproc && "
+        "chmod 666 /usr/local/share/pproc/whitelist.txt'"
+    );
+    
+    if (system(command) != 0) {
+        log_message(LL_ERROR, "Failed to setup whitelist directory and permissions");
+        return 1;
+    }
+
+    // Append to whitelist using echo and sudo
+    snprintf(command, sizeof(command), 
+        "echo '%s' | sudo tee -a /usr/local/share/pproc/whitelist.txt > /dev/null", 
+        file_path
+    );
+    
+    if (system(command) != 0) {
+        log_message(LL_ERROR, "Failed to add entry to whitelist");
+        return 1;
+    }
+
+    log_message(LL_INFO, "Added %s to whitelist", file_path);
+}
 
 int get_user_input(char *prompt)
 {
