+++ src/Utils/scanner.c
@@ -6,8 +6,6 @@
 #include <dirent.h>
 #include <unistd.h>
 #include <linux/limits.h>
-#include <errno.h>
-#include <linux/limits.h>
 
 #include "Crypto/fingerprint.h"
 #include "Utils/logger.h"
@@ -15,15 +13,8 @@
 #include "scanner.h"
 
 #define MAX_THREADS 10 // Max number of concurrent threads to avoid overload
-// Path to the whitelist file
-#define WHITELIST_PATH "/usr/local/share/pproc/whitelist.txt"
 
-// Mutex for synchronization
-pthread_mutex_t mutex = PTHREAD_MUTEX_INITIALIZER;
-
-pthread_mutex_t thread_count_mutex = PTHREAD_MUTEX_INITIALIZER;
-
-pthread_mutex_t whitelist_mutex = PTHREAD_MUTEX_INITIALIZER;
+pthread_mutex_t file_handler_mutex = PTHREAD_MUTEX_INITIALIZER;
 
 // Thread data structure to hold the file or directory path
 typedef struct
@@ -47,69 +38,6 @@ int active_threads = 0;
  */
 int scan_hashes(char *target_hash, char *target_file, char *hash_file, unsigned int hash_buffer_size, int automated_mode);
 
-int is_whitelisted(const char* target_path) {
-    int result = 0;
-    pthread_mutex_lock(&whitelist_mutex);
-    
-    FILE* whitelist_file = fopen(WHITELIST_PATH, "r");
-    if (whitelist_file == NULL) {
-        log_message(LL_DEBUG, "Whitelist file not found, assuming file is not whitelisted");
-        pthread_mutex_unlock(&whitelist_mutex);
-        return 0;
-    }
-
-    char line[PATH_MAX];
-    while (fgets(line, sizeof(line), whitelist_file)) {
-        // Remove newline character
-        line[strcspn(line, "\n")] = 0;
-        
-        if (strcmp(line, target_path) == 0) {
-            result = 1;
-            break;
-        }
-    }
-
-    fclose(whitelist_file);
-    pthread_mutex_unlock(&whitelist_mutex);
-    return result;
-}
-
-void add_to_whitelist(const char* file_path) {
-    pthread_mutex_lock(&whitelist_mutex);
-    
-    char command[PATH_MAX * 2];
-    
-    // Create directory and set permissions
-    snprintf(command, sizeof(command), 
-        "sudo sh -c '"
-        "mkdir -p /usr/local/share/pproc && "
-        "touch /usr/local/share/pproc/whitelist.txt && "
-        "chmod 777 /usr/local/share/pproc && "
-        "chmod 666 /usr/local/share/pproc/whitelist.txt'"
-    );
-    
-    if (system(command) != 0) {
-        log_message(LL_ERROR, "Failed to setup whitelist directory and permissions");
-        pthread_mutex_unlock(&whitelist_mutex);
-        return;
-    }
-
-    // Append to whitelist using echo and sudo
-    snprintf(command, sizeof(command), 
-        "echo '%s' | sudo tee -a /usr/local/share/pproc/whitelist.txt > /dev/null", 
-        file_path
-    );
-    
-    if (system(command) != 0) {
-        log_message(LL_ERROR, "Failed to add entry to whitelist");
-        pthread_mutex_unlock(&whitelist_mutex);
-        return;
-    }
-
-    log_message(LL_INFO, "Added %s to whitelist", file_path);
-    pthread_mutex_unlock(&whitelist_mutex);
-}
-
 int scan_file(char *target_file, int automated_mode)
 {
     // Check whitelist before scanning
@@ -122,6 +50,8 @@ int scan_file(char *target_file, int automated_mode)
     char target_sha1_hash[SHA1_BUFFER_SIZE];
     char target_sha256_hash[SHA256_BUFFER_SIZE];
     char target_md5_hash[MD5_BUFFER_SIZE];
+
+    //hold result of scan
     int scan_result;
 
     log_message(LL_INFO, "Starting File Scan for %s", target_file);
@@ -135,6 +65,7 @@ int scan_file(char *target_file, int automated_mode)
     }
 
     scan_result = scan_hashes(target_sha1_hash, target_file, "/usr/local/share/pproc/sha1-hashes.txt", SHA1_BUFFER_SIZE, automated_mode);
+
     if (scan_result == 1) {
         log_message(LL_WARNING, "Malicious file detected (SHA1) in %s", target_file);
         return 0;
@@ -155,6 +86,7 @@ int scan_file(char *target_file, int automated_mode)
     }
 
     scan_result = scan_hashes(target_sha256_hash, target_file, "/usr/local/share/pproc/sha256-hashes.txt", SHA256_BUFFER_SIZE, automated_mode);
+
     if (scan_result == 1)
     {
         log_message(LL_WARNING, "Malicious file detected (SHA256) in %s", target_file);
@@ -350,126 +282,20 @@ int scan_hashes(char *target_hash, char *target_file, char *hash_file, unsigned 
 
     while (fgets(current_hash, hash_buffer_size, hashes)) {
         if (strcmp(current_hash, target_hash) == 0) {
-            log_message(LL_WARNING, "Malicious file detected: %s", target_file);
-            
-            struct stat file_stat;
-            unsigned int file_permissions;
+            //free resources we found out the file is bad
+            free(current_hash);
+            fclose(hashes);
 
-            if (stat(target_file, &file_stat) == -1) {
-                log_message(LL_ERROR, "Could not stat file %s", target_file);
-                free(current_hash);
-                fclose(hashes);
+            //handle the file only one thread should do this at a time 
+            pthread_mutex_lock(&file_handler_mutex);
+
+            if (handle_file(target_file) != 0) {
+                log_message(LL_ERROR, "Could not proprely deal with file");
                 return -1;
             }
 
-            file_permissions = file_stat.st_mode;
-
-            if (chmod(target_file, S_IRUSR | S_IRGRP | S_IROTH) == -1) {
-                log_message(LL_ERROR, "Failed to change file permissions");
-                free(current_hash);
-                fclose(hashes);
-                return 1;
-            }
-
-            if (automated_mode) {
-                log_message(LL_INFO, "Automated mode: Removing malicious file %s", target_file);
-                if (remove(target_file) != 0) {
-                    log_message(LL_ERROR, "Failed to remove file in automated mode: %s", strerror(errno));
-                }
-            } else {
-                printf("\nPossible malicious file detected: %s", target_file);
-                if (get_user_input("\nWould you like to remove the file Y/N:") == 1) {
-                    log_message(LL_INFO, "Removing file %s", target_file);
-                    if (remove(target_file) != 0) {
-                        log_message(LL_ERROR, "Failed to remove file: %s", strerror(errno));
-                    }
-                } else {
-                    if (get_user_input("\nWould you like to quarantine the file Y/N:") == 1) {
-                        struct stat st = {0};
-                        if (stat("/usr/local/share/pproc/quarantine", &st) == -1) {
-                            if (mkdir("/usr/local/share/pproc/quarantine", 0777) == -1) {
-                                log_message(LL_ERROR, "Failed to create quarantine directory: %s", strerror(errno));
-                                chmod(target_file, file_permissions);
-                                break;
-                            }
-                            chmod("/usr/local/share/pproc/quarantine", 0777);
-                        }
-
-                        const char *filename = strrchr(target_file, '/');
-                        filename = filename ? filename + 1 : target_file;
-                        
-                        char quarantine_path[PATH_MAX];
-                        snprintf(quarantine_path, sizeof(quarantine_path), 
-                                "/usr/local/share/pproc/quarantine/%s", filename);
-
-                        // Open source file
-                        FILE *src = fopen(target_file, "rb");
-                        if (!src) {
-                            log_message(LL_ERROR, "Failed to open source file for quarantine: %s", strerror(errno));
-                            chmod(target_file, file_permissions);
-                            break;
-                        }
-
-                        // Open destination file with explicit permissions
-                        FILE *dst = fopen(quarantine_path, "wb");
-                        if (!dst) {
-                            log_message(LL_ERROR, "Failed to open quarantine destination: %s", strerror(errno));
-                            fclose(src);
-                            chmod(target_file, file_permissions);
-                            break;
-                        }
-
-                        // Set permissions on the quarantine file
-                        chmod(quarantine_path, 0666);
-
-                        // Copy file contents
-                        char buf[8192];
-                        size_t size;
-                        int copy_success = 1;
-                        
-                        while ((size = fread(buf, 1, sizeof(buf), src)) > 0) {
-                            if (fwrite(buf, 1, size, dst) != size) {
-                                copy_success = 0;
-                                break;
-                            }
-                        }
-
-                        fclose(src);
-                        fclose(dst);
-
-                        if (copy_success) {
-                            // Log the original path and permissions before removing the file
-                            FILE *quarantine_log = fopen("/usr/local/share/pproc/quarantine_log.txt", "a");
-                            if (quarantine_log) {
-                                fprintf(quarantine_log, "%s %o\n", target_file, file_permissions);
-                                fclose(quarantine_log);
-                            } else {
-                                log_message(LL_ERROR, "Failed to open quarantine log file");
-                            }
-
-                            if (remove(target_file) == 0) {
-                                log_message(LL_INFO, "File quarantined successfully: %s", filename);
-                            } else {
-                                log_message(LL_ERROR, "Failed to remove original file after quarantine: %s", strerror(errno));
-                                chmod(target_file, file_permissions);
-                                remove(quarantine_path);
-                            }
-                        } else {
-                            log_message(LL_ERROR, "Failed to copy file to quarantine: %s", strerror(errno));
-                            chmod(target_file, file_permissions);
-                            remove(quarantine_path);
-                        }
-                    } else {
-                        if (get_user_input("\nWould you like to add this file to the whitelist Y/N:") == 1) {
-                            add_to_whitelist(target_file);
-                        }
-                        chmod(target_file, file_permissions);
-                    }
-                }
-            }
+            pthread_mutex_unlock(&file_handler_mutex);
             
-            free(current_hash);
-            fclose(hashes);
             return 1;
         }
     }
